+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Following VS Code best practices, F1 uses semantic versioning for all releases.
 
 ---
 ## [2.1.1] - 2026-07-06
+- **Add**: Button cancel in "install-skill"
 - **Better**: filter chat
 - **REFACTOR**: .pe